refactor: category 관련 타이핑 추가
