--- src/types/sortOption.ts
@@ -1,3 +1,0 @@
-import { SORT_OPTIONS } from "../constants/products";
-
-export type SortOption = (typeof SORT_OPTIONS)[keyof typeof SORT_OPTIONS];
